requirements.txt
Flask
pandas
python-dotenv
psycopg2-binary
gunicorn
requests
isodate